CRAN residuals: generics re-export, portable include path, size guards
- tidy / glance re-export the generics-package generics instead of
  defining competitors (attaching tulpa next to broom no longer masks);
  generics added to Imports.
- src/Makevars moves -I../inst/include to PKG_CPPFLAGS (preprocessor
  flag, matching Makevars.win); cran-comments dontrun justification now
  also covers examples needing the off-CRAN tulpaRatio families.
- Size arithmetic hardening: the NUTS result copy and the NNGP workspace
  compute allocation sizes in size_t; the store_Q flat offsets (int in
  the exported ABI struct) accumulate in 64-bit and reject a wrapping
  grid loudly. shim_guard.h documents the longjmp caveat for consumer
  frames. Rd regenerated (k_samples rename, fitted/return contracts,
  reexports).

+++ src/tulpa_init.cpp
@@ -29,7 +29,9 @@ static void fill_nuts_result_from_cpp(
     out->sampler[63] = '\0';
 
     int ns = hmc.n_sample;
-    out->samples = new double[ns * n_params];
+    // size_t products: int * int overflows (UB) before ~2.1e9 elements, which
+    // a large-latent multi-chain run can reach.
+    out->samples = new double[static_cast<std::size_t>(ns) * n_params];
     out->log_prob = new double[ns];
     out->accept_prob = new double[ns];
     out->divergent = new int[ns];
@@ -38,7 +40,7 @@ static void fill_nuts_result_from_cpp(
     for (int s = 0; s < ns; s++) {
         const double* row = hmc.sample_row(s);
         for (int j = 0; j < n_params; j++) {
-            out->samples[s * n_params + j] = row[j];
+            out->samples[static_cast<std::size_t>(s) * n_params + j] = row[j];
         }
         out->log_prob[s] = hmc.log_prob[s];
         out->accept_prob[s] = hmc.accept_prob[s];
